+++ src/main/java/racingcar/ConsoleApplication.java
@@ -1,17 +1,17 @@
 package racingcar;
 
-import racingcar.controller.RacingGameController;
+import racingcar.controller.ConsoleController;
 import racingcar.domain.movingstrategy.DefaultMovingStrategy;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
 import java.util.Scanner;
 
-public final class Application {
+public final class ConsoleApplication {
 
     public static void main(String[] args) {
         try (final Scanner scanner = new Scanner(System.in)) {
-            RacingGameController controller = new RacingGameController(
+            ConsoleController controller = new ConsoleController(
                     new InputView(scanner), new OutputView());
             controller.run(new DefaultMovingStrategy());
         }
