+++ src/main/java/racingcar/controller/WebController.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -7,15 +7,18 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
+import racingcar.dto.GameResultResponseDto;
+import racingcar.dto.UserRequestDto;
+import racingcar.service.RacingGameService;
 
 import java.util.List;
 
 @Controller
-public class RacingCarController {
+public class WebController {
 
     private final RacingGameService racingGameService;
 
-    public RacingCarController(final RacingGameService racingGameService) {
+    public WebController(final RacingGameService racingGameService) {
         this.racingGameService = racingGameService;
     }
