+++ src/main/java/racingcar/dto/CarDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 public final class CarDto {
