+++ src/main/java/racingcar/dto/GameResultResponseDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 import racingcar.entity.CarEntity;
