+++ src/main/java/racingcar/dto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 public final class UserRequestDto {
