--- src/main/java/racingcar/view/exception/NotFoundViewException.java
@@ -1,4 +1,0 @@
-package racingcar.view.exception;
-
-public final class NotFoundViewException extends RuntimeException {
-}
